docs: adding phpdocs to salesforce job, relocating query builder to src/support
